requirements.txt
streamlit
pandas
scikit-learn
matplotlib
joblib
numpy
Pillow
requests